Make calculator into a component
Extracted calculator from adoption confirmation page, created
component. Did some refactoring.
