+++ src/app/components/calculator/calculator.hbs
@@ -1,0 +1,56 @@
+<table id="savings-calculator">
+    <thead>
+        <tr>
+            <th></th>
+            <th>OpenStax Prices</th>
+            <th>Traditional Prices</th>
+            <th>Number of Students</th>
+            <th>Savings Over Three Years</th>
+        </tr>
+    </thead>
+    <tbody>
+        <tr>
+            <th>E-book Only</th>
+            <td>$0.00</td>
+            <td>$<input type="number" value="100.00" /></td>
+            <td><input id="student-count" type="number" value="1000" /></td>
+            <td class="savings"></td>
+        </tr>
+        <tr>
+            <th>Printed Text Only</th>
+            <td>$<input type="number" value="41.00" /></td>
+            <td>$<input type="number" value="220.00"/></td>
+            <td>1000</td>
+            <td class="savings footnote"></td>
+        </tr>
+        <tr>
+            <th>Printed Text + Online Resources</th>
+            <td>$<input type="number" value="91.00" /></td>
+            <td>$<input type="number" value="245.00" /></td>
+            <td>1000</td>
+            <td class="savings"></td>
+        </tr>
+        <tr>
+            <th>E-book + Online Resources</th>
+            <td>$<input type="number" value="50.00" /></td>
+            <td>$<input type="number" value="145.00" /></td>
+            <td>1000</td>
+            <td class="savings"></td>
+        </tr>
+    </tbody>
+    <tfoot>
+        <tr>
+            <td colspan="5">
+                <p>
+                    Use as a guide; price varies widely by source and local markets.
+                    OpenStax College printed cost based on on-demand printing of
+                    <i>College Physics</i>
+                </p>
+                <p>
+                    <sup>*</sup> Assumes students will be able to purchase
+                    used copies, 10% used in year 1, 50% in year2, 75% in year 3.
+                </p>
+            </td>
+        </tr>
+    </tfoot>
+</table>
